+++ lib/services/impl/notifications_permission_service.dart
@@ -1,7 +1,8 @@
 import 'package:permission_handler/permission_handler.dart';
 import 'package:simple_permission_workflow/services/permission_service.dart';
 
-class SPWNotificationsPermission implements SPWPermissionService {
+class SPWNotificationsPermission
+    extends SPWPermissionService<SPWNotificationsPermission> {
   @override
   Future<PermissionStatus> request() async {
     return await Permission.notification.request();